Do not deal with deleted stop_times when checking consistency
The logic is the same for both arrival and departure, but as these are not contained in the same object, mutualization is a bit complicated.

Also, here we postpone everything inconsistent using the delay from previous arrival/departure-time... Which changes a bit the behavior when comparing departure and arrival at the same stop.
Note that before this consistency pass, merge() function already handles part of the things (which is not that good but I do not want to touch that now).
So this case should never happen here in practice.